components/services/microservices/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.2.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pika = "^1.3.2"
jwt = "^1.3.1"
peewee = "^3.17.8"
pyjwt = "^2.10.0"
psycopg2-binary = "^2.9.10"
service-runtime = { path = "../../service-runtime", develop = true }
argon2-cffi = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
